Updated game view

+++ app/src/main/kotlin/games/planetwars/agents/GameInterfaces.kt
@@ -14,14 +14,14 @@ data class Action (
     val playerId: Player, // the player that is making the move, set to Neutral to do nothing
     val sourcePlanetId: Int,
     val destinationPlanetId: Int,
-    val numShips: Int
+    val numShips: Double
 ) {
     companion object {
         val DO_NOTHING = Action(
             playerId = Player.Neutral,
             sourcePlanetId = -1,
             destinationPlanetId = -1,
-            numShips = 0
+            numShips = 0.0
         )
 
         fun doNothing(): Action = DO_NOTHING
